Refactoring and cleaning up done.
